fix: made page show properly and removed errors in the console
