+++ package.json
@@ -12,7 +12,9 @@
     "lint:fix": "next lint --fix"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.7",
     "axios": "^1.5.1",
+    "bufferutil": "^4.0.8",
     "express": "^4.18.2",
     "history": "^5.3.0",
     "moment": "^2.29.4",
@@ -29,7 +31,8 @@
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.2",
     "sass": "^1.69.4",
-    "socket.io-client": "^4.7.2"
+    "socket.io-client": "^4.7.2",
+    "utf-8-validate": "^6.0.3"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.22.15",
