+++ src/app/layout.tsx
@@ -1,3 +1,5 @@
+import StoreProvider from "@/store"
+
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
@@ -10,7 +12,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body><StoreProvider>{children}</StoreProvider></body>
     </html>
   )
 }
