+++ src/app/page.tsx
@@ -1,34 +1,47 @@
 'use client'
 import React from 'react';
-import { Provider } from 'react-redux';
 import 'react-toastify/dist/ReactToastify.css';
 import 'moment/locale/fr';
 import './page.scss';
-
-import Index from '../routes';
-import Sell from '../routes/sell';
-
-import store from '../store';
-import Tv from '../routes/tv';
-import Preparation from '../routes/preparation';
-import Items from '../routes/items';
-import LoginRouter from '../components/loginRouter';
-import { Route } from 'react-router';
 import moment from 'moment';
+import { useDispatch } from 'react-redux';
+import Navbar from '@/components/navbar';
+import FontAwesome from 'react-fontawesome';
+import { logout } from '@/reducers/login';
+import { Action } from 'redux';
 
 moment.locale('fr');
 
 const App = () => {
+  const dispatch = useDispatch();
   return (
-    <Provider store={store}>
-      <LoginRouter>
-        <Route exact path="/" Component={Index} />
-        <Route path="/sell" Component={Sell} />
-        <Route path="/preparation" Component={Preparation} />
-        <Route path="/tv" Component={Tv} />
-        <Route path="/items" Component={Items} />
-      </LoginRouter>
-    </Provider>
+    <>
+      <Navbar>
+        <div className="logout" onClick={() => dispatch(logout() as unknown as Action)}>
+          <FontAwesome name="sign-out-alt" /> Déconnexion
+        </div>
+      </Navbar>
+      {/* <div id="index">
+        <div onClick={() => history.push('/sell?except=goodies')}>
+          <FontAwesome name="hamburger" /> Vente de bouffe
+        </div>
+        <div onClick={() => history.push('/sell?only=goodies')}>
+          <FontAwesome name="tshirt" /> Vente de goodies
+        </div>
+        <div onClick={() => history.push('/preparation')}>
+          <FontAwesome name="check" /> Préparation générale
+        </div>
+        <div onClick={() => history.push('/preparation?only=pizzas')}>
+          <FontAwesome name="pizza-slice" /> Préparation des pizzas
+        </div>
+        <div onClick={() => history.push('/tv')}>
+          <FontAwesome name="tv" /> TV
+        </div>
+        <div onClick={() => history.push('/items')}>
+          <FontAwesome name="receipt" /> Gestion des items
+        </div>
+      </div> */}
+    </>
   );
 };
