+++ src/components/navbar.tsx
@@ -2,10 +2,10 @@ import React, { useEffect, useState } from 'react';
 import FontAwesome from 'react-fontawesome';
 import moment from 'moment';
 
-import { history } from '../components/loginRouter';
 import './navbar.scss';
 import { State } from '../types';
 import { useSelector } from 'react-redux';
+import { useRouter } from 'next/navigation';
 
 interface PropTypes {
   back?: string;
@@ -13,10 +13,11 @@ interface PropTypes {
   children?: React.ReactNode;
 }
 
-const Navbar = ({ back, onBack, children }: PropTypes) => {
+const Navbar = ({ back = null, onBack = null, children = null }: PropTypes) => {
   const [time, setTime] = useState(moment().format('H[h]mm'));
   const name = useSelector((state: State) => state.login.name);
   const serverOnline = useSelector((state: State) => state.server.internetConnected);
+  const router = useRouter();
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -32,7 +33,7 @@ const Navbar = ({ back, onBack, children }: PropTypes) => {
     }
 
     if (back) {
-      history.push(back);
+      router.push(back);
     }
   };
 
@@ -52,12 +53,6 @@ const Navbar = ({ back, onBack, children }: PropTypes) => {
       <div className="absolute-right">{children ? children : ''}</div>
     </nav>
   );
-};
-
-Navbar.defaultProps = {
-  back: null,
-  onBack: null,
-  children: null,
 };
 
 export default Navbar;
