+++ src/routes/tv.tsx
@@ -3,9 +3,9 @@ import React, { useEffect, useRef, LegacyRef, useState } from 'react';
 import './tv.scss';
 import { useSelector } from 'react-redux';
 
-import { history } from '../components/loginRouter';
 import { State, Order as OrderType, Status, OrderItem } from '../types';
 import Separator from '../components/UI/separator';
+import { useRouter } from 'next/navigation';
 
 interface GroupedCategory {
   name: string;
@@ -74,6 +74,7 @@ const Tv = () => {
   const readyOrders = orders.filter((order) => order.status === Status.READY);
 
   const refs = useRef<HTMLDivElement[]>([null, null, null]);
+  const router = useRouter();
 
   const speedDown = 2;
   const speedUp = -5;
@@ -109,7 +110,7 @@ const Tv = () => {
   });
 
   return (
-    <div id="tv" onClick={() => history.push('/')}>
+    <div id="tv" onClick={() => router.push('/')}>
       <div className="orders">
         <div className="title">En attente</div>
         <OrderGrid orders={pendingOrders} passingRef={(el) => (refs.current[0] = el)} />
