--- src/store.ts
@@ -1,9 +1,0 @@
-import { createLogger } from 'redux-logger';
-import { applyMiddleware, createStore } from 'redux';
-import thunk from 'redux-thunk';
-import reducers from './reducers';
-
-const logger = createLogger({ collapsed: true });
-const store = createStore(reducers, applyMiddleware(thunk, logger));
-
-export default store;
